III-3137 Move history services to separate providers
